+++ src/content/dto/content-summary.dto.ts
