+++ src/content/entities/content.entity.ts
@@ -1,7 +1,7 @@
-import { ContentSummaryDto } from 'src/collection/dto/response/content-summary.dto';
 import { BaseEntity } from 'src/common/entities/base.entity';
 import { ContentType } from 'src/common/enums/content-type.enum';
 import { SourceType } from 'src/common/enums/source-type.enum';
+import { ContentSummaryDto } from 'src/content/dto/content-summary.dto';
 import { Column, Entity, OneToMany, TableInheritance } from 'typeorm';
 import { ContentGenre } from './content-genre.entity';
