+++ src/content/movie/entities/movie.entity.ts
@@ -1,4 +1,4 @@
-import { ContentSummaryDto } from 'src/collection/dto/response/content-summary.dto';
+import { ContentSummaryDto } from 'src/content/dto/content-summary.dto';
 import { Content } from 'src/content/entities/content.entity';
 import { MovieStatus } from 'src/content/enum/movie-status.enum';
 import { ChildEntity, Column } from 'typeorm';
