+++ src/content/tv-series/entities/tv-series.entity.ts
@@ -1,4 +1,4 @@
-import { ContentSummaryDto } from 'src/collection/dto/response/content-summary.dto';
+import { ContentSummaryDto } from 'src/content/dto/content-summary.dto';
 import { Content } from 'src/content/entities/content.entity';
 import { TvSeriesStatus } from 'src/content/enum/tv-series-status.enum';
 import { ChildEntity, Column, OneToMany } from 'typeorm';
